* Clean up and docs fixes by exposing Python APIs * Restructring 'python/tests' directory - restored 'compiler' directory   and fixed 'PYTHONPATH' in the run command.
