+++ .github/workflows/python_wheels.yml
@@ -231,6 +231,13 @@ jobs:
               exit 1
             fi
 
+            python${{ inputs.python_version }} -m pytest -v /tmp/tests/compiler
+            pytest_status=$?
+            if [ ! $pytest_status -eq 0 ]; then
+              echo "::error file=python_wheel.yml::Python 'compiler' tests failed with status $pytest_status."
+              exit 1
+            fi
+
             python${{ inputs.python_version }} -m pytest -v /tmp/tests/domains
             pytest_status=$?
             if [ ! $pytest_status -eq 0 ]; then
