+++ python/tests/compiler/float.py
@@ -1,12 +1,12 @@
 # ============================================================================ #
-# Copyright (c) 2022 - 2024 NVIDIA Corporation & Affiliates.                   #
+# Copyright (c) 2022 - 2023 NVIDIA Corporation & Affiliates.                   #
 # All rights reserved.                                                         #
 #                                                                              #
 # This source code and the accompanying materials are made available under     #
 # the terms of the Apache License 2.0 which accompanies this distribution.     #
 # ============================================================================ #
 
-# RUN: PYTHONPATH=../../.. pytest -rP  %s | FileCheck %s
+# RUN: PYTHONPATH=../../ pytest -rP  %s | FileCheck %s
 
 import os
 
@@ -16,11 +16,11 @@ import numpy as np
 import cudaq
 
 
-def test_make_kernel_list():
+def test_make_kernel_float():
     """
-    Test `cudaq.make_kernel` with a list of floats as parameters.
+    Test `cudaq.make_kernel` with one float parameter.
     """
-    kernel, parameter = cudaq.make_kernel(list)
+    kernel, parameter = cudaq.make_kernel(float)
     # Kernel should only have 1 argument and parameter.
     got_arguments = kernel.arguments
     got_argument_count = kernel.argument_count
@@ -31,7 +31,7 @@ def test_make_kernel_list():
 
 
 # CHECK-LABEL:   func.func @__nvqpp__mlirgen____nvqppBuilderKernel_{{.*}}(
-# CHECK-SAME:    %[[VAL_0:.*]]: !cc.stdvec<f64>) attributes {"cudaq-entrypoint"} {
+# CHECK-SAME:    %[[VAL_0:.*]]: f64) attributes {"cudaq-entrypoint"} {
 # CHECK:           return
 # CHECK:         }
