+++ python/tests/compiler/lit.cfg.py
@@ -1,5 +1,5 @@
 # ============================================================================ #
-# Copyright (c) 2022 - 2024 NVIDIA Corporation & Affiliates.                   #
+# Copyright (c) 2022 - 2023 NVIDIA Corporation & Affiliates.                   #
 # All rights reserved.                                                         #
 #                                                                              #
 # This source code and the accompanying materials are made available under     #
@@ -16,7 +16,7 @@ from lit.llvm import llvm_config
 # Configuration file for the 'lit' test runner.
 
 # name: The name of this test suite.
-config.name = 'pyCUDAQMLIR'
+config.name = 'pyCUDAQ'
 
 # testFormat: The test format to use to interpret tests.
 #
@@ -54,7 +54,7 @@ config.test_source_root = os.path.dirname(__file__)
 
 # test_exec_root: The root path where tests should be run.
 config.test_exec_root = os.path.join(config.cudaq_obj_root,
-                                     'python/tests/mlir')
+                                     'python/tests/compiler')
 
 # Tweak the PATH to include the tools dir.
 llvm_config.with_environment('PATH', config.cudaq_tools_dir, append_path=True)
