+++ python/tests/compiler/mixed_args.py
@@ -1,12 +1,12 @@
 # ============================================================================ #
-# Copyright (c) 2022 - 2024 NVIDIA Corporation & Affiliates.                   #
+# Copyright (c) 2022 - 2023 NVIDIA Corporation & Affiliates.                   #
 # All rights reserved.                                                         #
 #                                                                              #
 # This source code and the accompanying materials are made available under     #
 # the terms of the Apache License 2.0 which accompanies this distribution.     #
 # ============================================================================ #
 
-# RUN: PYTHONPATH=../../.. pytest -rP  %s | FileCheck %s
+# RUN: PYTHONPATH=../../ pytest -rP  %s | FileCheck %s
 
 import os
