+++ python/tests/compiler/multiple_floats.py
@@ -1,12 +1,12 @@
 # ============================================================================ #
-# Copyright (c) 2022 - 2024 NVIDIA Corporation & Affiliates.                   #
+# Copyright (c) 2022 - 2023 NVIDIA Corporation & Affiliates.                   #
 # All rights reserved.                                                         #
 #                                                                              #
 # This source code and the accompanying materials are made available under     #
 # the terms of the Apache License 2.0 which accompanies this distribution.     #
 # ============================================================================ #
 
-# RUN: PYTHONPATH=../../.. pytest -rP  %s | FileCheck %s
+# RUN: PYTHONPATH=../../ pytest -rP  %s | FileCheck %s
 
 import os
 
@@ -16,22 +16,22 @@ import numpy as np
 import cudaq
 
 
-def test_make_kernel_float():
+def test_make_kernel_multiple_floats():
     """
-    Test `cudaq.make_kernel` with one float parameter.
+    Test `cudaq.make_kernel` with multiple parameters.
     """
-    kernel, parameter = cudaq.make_kernel(float)
-    # Kernel should only have 1 argument and parameter.
+    kernel, parameter_1, parameter_2 = cudaq.make_kernel(float, float)
+    # Kernel should have 2 arguments and parameters.
     got_arguments = kernel.arguments
     got_argument_count = kernel.argument_count
-    assert len(got_arguments) == 1
-    assert got_argument_count == 1
-    # Dump the MLIR for FileCheck.
+    assert len(got_arguments) == 2
+    assert got_argument_count == 2
     print(kernel)
 
 
 # CHECK-LABEL:   func.func @__nvqpp__mlirgen____nvqppBuilderKernel_{{.*}}(
-# CHECK-SAME:    %[[VAL_0:.*]]: f64) attributes {"cudaq-entrypoint"} {
+# CHECK-SAME:    %[[VAL_0:.*]]: f64,
+# CHECK:         %[[VAL_1:.*]]: f64) attributes {"cudaq-entrypoint"} {
 # CHECK:           return
 # CHECK:         }
