+++ python/tests/mlir/ast_list_int.py
@@ -6,7 +6,7 @@
 # the terms of the Apache License 2.0 which accompanies this distribution.     #
 # ============================================================================ #
 
-# RUN: PYTHONPATH=../../../.. pytest -rP  %s | FileCheck %s
+# RUN: PYTHONPATH=../../ pytest -rP  %s | FileCheck %s
 
 import os
